new project structure + PEP-8 style file names
